2.1-databases/work_with_database/requirements.txt
django~=5.1.4
psycopg2-binary

django-autoslug~=1.9.9